quick fix parakeet download

+++ backend/utils/voice/download_models_script.py
@@ -26,14 +26,15 @@ if not os.path.isfile(SILERO_TTS_MODEL_PATH):
 
 #parakeet
 def download_file(url, local_path):
-    response = requests.get(url, stream=True)
-    response.raise_for_status()
-    
-    os.makedirs(os.path.dirname(local_path), exist_ok=True)
-    
-    with open(local_path, 'wb') as f:
-        for chunk in response.iter_content(chunk_size=8192):
-            f.write(chunk)
+    if not os.path.isfile(local_path):
+        response = requests.get(url, stream=True)
+        response.raise_for_status()
+        
+        os.makedirs(os.path.dirname(local_path), exist_ok=True)
+        
+        with open(local_path, 'wb') as f:
+            for chunk in response.iter_content(chunk_size=8192):
+                f.write(chunk)
 
 
 repo_id = "istupakov/parakeet-tdt-0.6b-v3-onnx"
